Code/BiBEx/BiBEx_Model/requirements.txt
fastapi==0.88.0
pydantic==1.10.2
uvicorn==0.20.0
python-multipart==0.0.5
transformers==4.25.1
colorama==0.4.6
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
tokenizers==0.13.2
pytesseract==0.3.10
PyMuPDF==1.21.1
Pillow==9.3.0
dehyphen==0.3.4